Fix wrong data type in Windows function override (STR #3232).
This patch uses the correct data type (as defined by Windows / msn)
but does not fix the Cygwin linker problem. Unfortunately.

This issue may be a Cygwin problem and should be reported to Cygwin.


git-svn-id: file:///fltk/svn/fltk/branches/branch-1.3@10795 ea41ed52-d2ee-0310-a9c1-e6b18d33e121

+++ src/fl_dnd_win32.cxx
@@ -334,7 +334,7 @@ public:
       delete this;
     return nTemp;
   }
-  STDMETHODIMP GiveFeedback( ulong ) { return DRAGDROP_S_USEDEFAULTCURSORS; }
+  STDMETHODIMP GiveFeedback( DWORD ) { return DRAGDROP_S_USEDEFAULTCURSORS; }
   STDMETHODIMP QueryContinueDrag( BOOL esc, DWORD keyState ) {
     if ( esc )
       return DRAGDROP_S_CANCEL;
